backend/requirements.txt
fastapi
uvicorn
faiss-cpu
qdrant-client
numpy
pandas
pydantic
python-multipart
httpx
python-dotenv
schedule
psutil
pymongo
redis
motor
grpcio
grpcio-tools
protobuf
torchserve
torch-model-archiver
torch-workflow-archiver
celery
flower
prometheus-client
structlog